move remapParameterValuesToTemplateTags
