+++ frontend/src/metabase-lib/lib/Question.ts
@@ -29,10 +29,7 @@ import { memoizeClass, sortObject } from "metabase-lib/lib/utils";
 
 // TODO: remove these dependencies
 import * as Urls from "metabase/lib/urls";
-import {
-  getCardUiParameters,
-  remapParameterValuesToTemplateTags,
-} from "metabase/parameters/utils/cards";
+import { getCardUiParameters } from "metabase/parameters/utils/cards";
 import {
   normalizeParameterValue,
   getParameterValuesBySlug,
@@ -60,6 +57,7 @@ import { DependentMetadataItem } from "metabase-types/types/Query";
 import { utf8_to_b64url } from "metabase/lib/encoding";
 import { CollectionId } from "metabase-types/api";
 
+import { remapParameterValuesToTemplateTags } from "metabase-lib/lib/parameters/utils/template-tags";
 import { fieldFilterParameterToMBQLFilter } from "metabase-lib/lib/parameters/utils/mbql";
 import { getQuestionVirtualTableId } from "metabase-lib/lib/metadata/utils/saved-questions";
 import {
